MyMenuList, Followers and Followed service request.

+++ cooktime1/src/main/java/com/cooktime/model/Recipe.java
@@ -181,7 +181,7 @@ public class Recipe {
         
     }
 
-    public float getCalification() {
+    public int getCalification() {
         
         return calification;
